changement quebec flag global
